% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Linear Regression function}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{a formula class object, similar to lm().}

\item{data}{input data frame.}
}
\value{
table similar to the coefficent table from summary() with rows for each coefficient
and columns for the Estimate, Std. Error, t value, and Pr(>|t|).
}
\description{
This function fits a linear model in R
}
\examples{
my_lm (mpg ~ hp + wt, data = mtcars)

}
\keyword{\code{inference}}
