% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t_test}
\alias{my_t_test}
\title{t-test function}
\usage{
my_t_test(x, alternative, mu)
}
\arguments{
\item{x}{a numeric vector of data.}

\item{alternative}{a character string specifying the alternative hypothesis.
This should only accept "two.sided", "less", or "greater".
Otherwise, the function should throw an informative error asking you to try again.}

\item{mu}{a number indicating the null hypothesis value of the mean.}
}
\value{
test_stat represents the numeric test statistic, df represents the degrees of freedom.
alternative represents the value of the parameter alternative, and p_val represents  the numeric p-value.
}
\description{
This function performs a one sample t-test in R.
}
\examples{
my_t_test(x = rnorm(100), alternative = "less", mu = 0 )
my_t_test(x = rnorm(100), alternative = "greater", mu = 0 )

}
\keyword{inference}
